feat(multisite): allow configuring entry points, head fields in multisite.config.js
