+++ packages/core/__mocks__/multisite.config.js
@@ -1,7 +1,7 @@
 module.exports = [
   {
     domain: 'multisite-one.irving.test',
-    vars: {
+    env: {
       API_ROOT_URL: 'https://irving-multisite.test/api',
       ROOT_URL: 'https://multisite-one.irving.test',
       BASIC_AUTH_USERNAME: 'alley',
@@ -11,7 +11,7 @@ module.exports = [
   },
   {
     domain: 'multisite-two.irving.test',
-    vars: {
+    env: {
       API_ROOT_URL: 'https://irving-multisite-two.test/api',
       ROOT_URL: 'https://multisite-two.irving.test',
     },
