+++ packages/core/config/multisite/index.js
@@ -1,0 +1,9 @@
+const { getSiteConfig } = require('./getSiteConfig');
+const getEnv = require('./getEnv');
+const getEntries = require('./getEntries');
+
+module.exports = {
+  getSiteConfig,
+  getEnv,
+  getEntries,
+};
