+++ packages/core/server/cache.js
@@ -3,7 +3,7 @@ const cors = require('cors');
 const purgeCache = require('./purgeCache');
 const getCacheKeys = require('./getCacheKeys');
 const createCheckAuth = require('./auth');
-const { getEnv } = require('../config/irving/multisite');
+const { getEnv } = require('../config/multisite');
 
 // Necessary to wrap this to get the request (and hostname) in scope.
 const corsMiddleware = (req, res, next) => {
