+++ packages/core/services/fetchComponentData.js
@@ -2,7 +2,7 @@ import AbortController from 'abort-controller';
 import omit from 'lodash/fp/omit';
 import getLogService from '@irvingjs/services/logService';
 import getCacheService from '@irvingjs/services/cacheService';
-import { getEnv } from 'config/irving/multisite';
+import { getEnv } from 'config/multisite';
 
 const log = getLogService('irving:components:data');
